json: switch to using ints in parse state for integral components of numbers
This introduces a spec-compliance issue: leading zeros in numbers
are now accepted, whereas the spec disallows them.  The reasoning
is that it is okay to be slightly more lenient than the spec,
especially if there is a performance gain.

There are still lists involved in handling fractional and exponent
components of numbers.